# clients/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/object_info_by_accessions": {
      "get": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Get Object Info By Accession",
        "operationId": "get_object_info_by_accession",
        "parameters": [
          {
            "name": "accessions",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "title": "Accessions"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ObjectInfo"
                  },
                  "title": "Response Get Object Info By Accession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/studies/{study_accession}/images_by_aliases": {
      "get": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Get Study Images By Alias",
        "operationId": "get_study_images_by_alias",
        "parameters": [
          {
            "name": "study_accession",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Study Accession"
            }
          },
          {
            "name": "aliases",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "title": "Aliases"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BIAImage"
                  },
                  "title": "Response Get Study Images By Alias"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/studies/{study_uuid}": {
      "get": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Get Study",
        "operationId": "get_study",
        "parameters": [
          {
            "name": "study_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Study Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BIAStudy"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/studies/{study_uuid}/file_references": {
      "get": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Get Study File References",
        "description": "First item in response is the next item with uuid greater than start_uuid.\nstart_uuid is part of the response",
        "operationId": "get_study_file_references",
        "parameters": [
          {
            "name": "study_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Study Uuid"
            }
          },
          {
            "name": "start_uuid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileReference"
                  },
                  "title": "Response Get Study File References"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/studies": {
      "get": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Search Studies",
        "description": "@TODO: Define search criteria for the general case\n\nFirst item in response is the next item with uuid greater than start_uuid.\nstart_uuid is part of the response",
        "operationId": "search_studies",
        "parameters": [
          {
            "name": "start_uuid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BIAStudy"
                  },
                  "title": "Response Search Studies"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/images/exact_match": {
      "post": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Search Images Exact Match",
        "description": "Exact match search of images with a specific attribute.\nMultiple parameters mean AND (as in, p1 AND p2).\nItems in lists with the `_any` suffix are ORed.\n\nAlthough `study_uuid` is optional, passing it if known is highly recommended and results in faster queries. Queries time out after 2 seconds, which should be enough for any search filtered by study.\n\nThis is likely to change fast, so **named arguments are recommended** in client apps instead of positional if possible to prevent downstream breakage.",
        "operationId": "search_images_exact_match",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchImageFilter"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BIAImage"
                  },
                  "type": "array",
                  "title": "Response Search Images Exact Match"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/file_references/exact_match": {
      "post": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Search File References Exact Match",
        "description": "Exact match search of file references with a specific attribute.\nMultiple parameters mean AND (as in, p1 AND p2).\nItems in lists with the `_any` suffix are ORed.\n\nAlthough `study_uuid` is optional, passing it if known is highly recommended and results in faster queries. Queries time out after 2 seconds, which should be enough for any search filtered by study.\n\nThis is likely to change fast, so **named arguments are recommended** in client apps instead of positional if possible to prevent downstream breakage.",
        "operationId": "search_file_references_exact_match",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchFileReferenceFilter"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FileReference"
                  },
                  "type": "array",
                  "title": "Response Search File References Exact Match"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/studies/exact_match": {
      "post": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Search Studies Exact Match",
        "operationId": "search_studies_exact_match",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchStudyFilter"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BIAStudy"
                  },
                  "type": "array",
                  "title": "Response Search Studies Exact Match"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/studies/{study_uuid}/images": {
      "get": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Get Study Images",
        "description": "First item in response is the next item with uuid greater than start_uuid.\nstart_uuid is part of the response",
        "operationId": "get_study_images",
        "parameters": [
          {
            "name": "study_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Study Uuid"
            }
          },
          {
            "name": "start_uuid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BIAImage"
                  },
                  "title": "Response Get Study Images"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/{image_uuid}": {
      "get": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Get Image",
        "operationId": "get_image",
        "parameters": [
          {
            "name": "image_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BIAImage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/images/{image_uuid}/ome_metadata": {
      "get": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Get Image Ome Metadata",
        "operationId": "get_image_ome_metadata",
        "parameters": [
          {
            "name": "image_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BIAImageOmeMetadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/file_references/{file_reference_uuid}": {
      "get": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Get File Reference",
        "operationId": "get_file_reference",
        "parameters": [
          {
            "name": "file_reference_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Reference Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileReference"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections": {
      "get": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Search Collections",
        "operationId": "search_collections",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BIACollection"
                  },
                  "title": "Response Search Collections"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/{collection_uuid}": {
      "get": {
        "tags": [
          "public",
          "private"
        ],
        "summary": "Get Collection",
        "operationId": "get_collection",
        "parameters": [
          {
            "name": "collection_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Collection Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BIACollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FileReference": {
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Uuid"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "study_uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Study Uuid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "uri": {
            "type": "string",
            "title": "Uri"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "size_in_bytes": {
            "type": "integer",
            "title": "Size In Bytes"
          },
          "attributes": {
            "type": "object",
            "title": "Attributes",
            "default": {}
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "uuid",
          "version",
          "study_uuid",
          "name",
          "uri",
          "type",
          "size_in_bytes"
        ],
        "title": "FileReference",
        "description": "A reference to an externally hosted file."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelMetadata": {
        "properties": {
          "type_name": {
            "type": "string",
            "title": "Type Name"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "type_name",
          "version"
        ],
        "title": "ModelMetadata"
      },
      "ObjectInfo": {
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "title": "Uuid"
          },
          "model": {
            "$ref": "#/components/schemas/ModelMetadata"
          }
        },
        "type": "object",
        "required": [
          "uuid",
          "model"
        ],
        "title": "ObjectInfo"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}